% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{build_uri}
\alias{build_uri}
\title{build uri}
\usage{
mindat_api_endpoint(name,endpoint_base,...,default_uri_builder,c(fields))
}
\arguments{
\item{...}{Further named parameters, other conditions.}

\item{name}{string .}

\item{endpoint_base}{list.}

\item{uri_builder}{function default is the default_uri_builder.}

\item{query_params}{list.}
}
\value{
uri string
}
\description{
build a request uri based on .
}
\examples{
mindat_api_endpoint("geomaterials","geomaterials/\%s",default_uri_builder,c(''))
mindat_api_endpoint('minerals_ima_list', 'minerals_ima/\%s', uri_builder = mindat_uri_builder)
}
