% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minda_data_maker.R
\name{ConvertDF2JsonLD}
\alias{ConvertDF2JsonLD}
\title{Output file as a given format}
\usage{
ConvertDF2JsonLD(inputdata,template = NULL)
}
\arguments{
\item{inputdata}{R dataframe of retrieived data from Mindat database.}

\item{template}{filepath to the template}
}
\description{
Convert the mindat R dataframe to JSON-LD string
}
\examples{
\dontrun{
df <-geomaterials_search_name("Quartz")
df_out <-ConvertDF2JsonLD(df)
}
}
