% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minda_data_maker.R
\name{ConvertDF2TTL}
\alias{ConvertDF2TTL}
\title{Convert a dataframe to a string of TTL format}
\usage{
ConvertDF2TTL (DataFrame)
}
\arguments{
\item{inputdata}{R dataframe of retrieived data from Mindat database.}

\item{template}{filepath to the template}
}
\description{
Convert the mindat R dataframe to TTL string
}
