% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_crystalclasses.R
\name{crystalclasses_symbols}
\alias{crystalclasses_symbols}
\title{crystalclasses that match a given vector of symbols (case-insenstive)}
\usage{
crystalclasses_symbols(symbols, ...)
}
\arguments{
\item{symbols}{vector of given crystals (array of strings or null).
The field "symbol" is to describe the symbol of Crystal class dictionary.}

\item{..., }{Further parameters like "system"(crystal system) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of crystalclasses
}
\description{
: Queries a list of crystalclasses that match a given list of symbols
}
\details{
This function filter data by a given list of symbols of crystal class dictionary
case-insensitive
}
\examples{
\dontrun{
df <-crystalclasses_symbols(c("2/m","mm2"))
}
}
