% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_crystalclasses.R
\name{crystalclasses_systems}
\alias{crystalclasses_systems}
\title{crystalclasses that match a given vector of crystal system (case-insenstive)}
\usage{
crystalclasses_systems(systems, ...)
}
\arguments{
\item{systems}{vector of given system.
"crystal system of the mineral;
"Amorphous","Hexagonal","Icosahedral","Isometric","Monoclinic","Orthorhombic","Tetragonal","Triclinic","Trigonal"}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of crystalclasses
}
\description{
: Queries a list of crystalclasses that match a given list of crystal system.
}
\details{
This function filter data by a given list of crystal_system of crystal class dictionary
case-insensitive
}
\examples{
\dontrun{
df <-geomaterials_crystal_system(c("Icosahedral"))
}
}
