% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_classifications.R
\name{Dana8_groups}
\alias{Dana8_groups}
\title{dana_8 classification}
\usage{
Dana8_groups(...)
}
\arguments{
\item{..., }{Further parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of dana8 groups
}
\description{
: Queries a list of Dana 8th edition classifications.
}
\details{
This function return a list of dana8 groups
case-insensitive
}
\examples{
\dontrun{
df <-Dana8_groups()
}
}
