% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_classifications.R
\name{Dana8_subgroups}
\alias{Dana8_subgroups}
\title{dana_8 subgroups}
\usage{
Dana8_subgroups(...)
}
\arguments{
\item{..., }{Further parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of dana8 subgroups
}
\description{
: Queries a list of subgroups of the Dana 8th edition classifications.
}
\details{
This function return a list of dana8 subgroups
case-insensitive
}
\examples{
\dontrun{
df <-Dana8_subgroups()
}
}
