% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_bi_greater_than}
\alias{geomaterials_bi_greater_than}
\title{retrieve the geomaterials whose birifrigence are higher than the given value.}
\usage{
geomaterials_bi_greater_than(gt, ...)
}
\arguments{
\item{gt}{float value.Birifrigence is calculated from refractive index as (rimax-rimin).Range: bi_min - bi_max.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that minmum value of the given birifrigence value.
}
\details{
This function releated to the field "bi_min" of geomaterials.
retrieve all the geomaterials that has higher birifrigence than the given value(gt).
}
\examples{
\dontrun{
df <-geomaterials_bi_greater_than(0.2)
}
}
