% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_bi_less_than}
\alias{geomaterials_bi_less_than}
\title{retrieve the geomaterials whose birifrigence are lower density than the given value.}
\usage{
geomaterials_bi_less_than(lt, ...)
}
\arguments{
\item{lt}{float value.Birifrigence is calculated from refractive index as (rimax-rimin).Range: bi_min - bi_max.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower birifrigence than lt.
}
\details{
This function releated to the field "bi_max" of geomaterials.
retrieve all the geomaterials that has higher birifrigence than the given value(lt).
}
\examples{
\dontrun{
df <-geomaterials_bi_less_than(0.3)
}
}
