% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_by_groupid}
\alias{geomaterials_by_groupid}
\title{retrieve the geomaterials by an given value of groupid.}
\usage{
geomaterials_by_groupid(id, ...)
}
\arguments{
\item{gid}{integer value. The id of the group to which this mineral belongs}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that match an given groupid.
}
\details{
This function releated to the field "groupid" of geomaterials.
retrieve all the geomaterials that match an given groupid.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_by_groupid(6)
}
