% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_cleavagetype}
\alias{geomaterials_cleavagetype}
\title{geomaterials that have the given cleavagetype}
\usage{
geomaterials_cleavagetype(types, ...)
}
\arguments{
\item{types}{vector of given cleavagetype}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the specified cleavagetype
}
\details{
Items Enum: "Distinct/Good" "Imperfect/Fair" "None Observed" "Perfect" "Poor/Indistinct" "Very Good"
}
\examples{
geomaterials_cleavagetype(c('Hexagonal'))
}
