% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_colour}
\alias{geomaterials_colour}
\title{geomaterials that have the given colors}
\usage{
geomaterials_colour(colors, ...)
}
\arguments{
\item{colors}{vector of given colors}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the specified colors
}
\details{
such: "Brown" "Yellow" "green"
}
\examples{
geomaterials_colour(c('Yellow'))
}
