% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_contain_all_and_without_elems}
\alias{geomaterials_contain_all_and_without_elems}
\title{geomaterials_that do not contain the elements}
\usage{
geomaterials_contain_all_and_without_elems(icl_elm_vector, ecl_elms_vector, ...)
}
\arguments{
\item{icl_elm_vector}{vector of elements.}

\item{ecl_elms_vector}{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
Queries the list of geological materials that simultaneously contain the specified elements and do not contain the specified elements.
}
\details{
This function queries the list of geological materials that do not contain the specified elements.
It performs the query operation by calling the mindat_geomaterial_list function.
}
\examples{
geomaterials_contain_all_and_without_elems(c('H','Be'),c('O'))
}
