% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_contain_all_elems}
\alias{geomaterials_contain_all_elems}
\title{geomaterials_contain_all_elems}
\usage{
geomaterials_contain_all_elems (icl_elms_vector)
}
\arguments{
\item{icl_elms_vector, }{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments that can be
passed to the mindat_geomaterial_list function.}
}
\value{
df, a data frame of geomaterials list.
}
\description{
retrieve the geomaterials that contain all of the elements.This function queries
the list of geomaterials that contain all the given elements.
It performs the query operation by calling the mindat_geomaterial_list function
}
\details{
This function releated to the field "elements_inc" of geomaterials.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_contain_all_elems(c('Fe','S'))
}
