% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_contain_any_elems}
\alias{geomaterials_contain_any_elems}
\title{geomaterials that contain any one of the given elements}
\usage{
geomaterials_contain_any_elems(any_elems, ...)
}
\arguments{
\item{any_elems}{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials.
}
\description{
: Queries the list of geological materials that contain any one of the given elements.
}
\details{
This function releated to the field "elements_inc" of geomaterials.
This function queries the list of geological materials that contain any element of an given list (any_elems).
It performs the query operation by looping through each given element and calling the mindat_geomaterial_list function.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_contain_any_elems(c('Dy','Be'))
}
