% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_dens_greater_than}
\alias{geomaterials_dens_greater_than}
\title{retrieve the geomaterials whose density are higher density than the given value.}
\usage{
geomaterials_dens_greater_than(gt, ...)
}
\arguments{
\item{gt}{float value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have higher density than gt.
}
\details{
retrieve all the geomaterials that has higher density than the given density.
}
\examples{
geomaterials_dens_greater_than(6)
geomaterials_dens_greater_than(3)
}
