% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_dens_less_than}
\alias{geomaterials_dens_less_than}
\title{retrieve the geomaterials whose density are lower density than the given value.}
\usage{
geomaterials_dens_less_than(lt, ...)
}
\arguments{
\item{..., }{Further named parameters.Other optional arguments.}

\item{gt}{float value}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower density than lt.
}
\details{
retrieve all the geomaterials that has higher density than the given density.
}
\examples{
geomaterials_dens_less_than(6)
geomaterials_dens_less_than(3)
}
