% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_dens_range}
\alias{geomaterials_dens_range}
\title{retrieve the geomaterials whose density are higher and lower than the given value.}
\usage{
geomaterials_dens_range(gt,lt, ...)
}
\arguments{
\item{gt}{float value}

\item{lt}{float value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower density than lt.
}
\details{
retrieve all the geomaterials that has the density within the range of (gt,lt).
}
\examples{
geomaterials_dens_range(3,3.5)
}
