% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_diapheny}
\alias{geomaterials_diapheny}
\title{retrieve the geomaterials that have the given diapheny.}
\usage{
geomaterials_diapheny(diapheny, ...)
}
\arguments{
\item{diapheny}{string}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given diapheny.
}
\details{
Items Enum: "Opaque" "Translucent" "Transparent"
}
\examples{
geomaterials_diapheny("Opaque")
}
