% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_entrytype}
\alias{geomaterials_entrytype}
\title{retrieve the geomaterials that have the given entrytype}
\usage{
geomaterials_entrytype(type, ...)
}
\arguments{
\item{types}{list of entry types.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given entrytype
}
\details{
Items Enum: 0 1 2 3 4 5 6 7 8
}
\examples{
geomaterials_entrytype(c('1'))
geomaterials_entrytype(c('1','2'))
}
