% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_fracturetype}
\alias{geomaterials_fracturetype}
\title{retrieve the geomaterials that have the given fracturetype.}
\usage{
geomaterials_fracturetype(types, ...)
}
\arguments{
\item{types}{list of types.fracturetype(Array of strings or null):
How the mineral breaks-"Conchoidal" "Fibrous" "Hackly" "Irregular/Uneven" "Micaceous" "None observed" "Splintery" "Step-Like" "Sub-Conchoidal".}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given fracturetype.
}
\details{
This function releated to the field "fracturetype" of geomaterials.
fracturetype(Items Enum): "Conchoidal" "Fibrous" "Hackly" "Irregular/Uneven" "Micaceous" "None observed" "Splintery" "Step-Like" "Sub-Conchoidal"
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_fracturetype(c('Conchoidal','Fibrous'))
}
