% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_hardness_gt}
\alias{geomaterials_hardness_gt}
\title{retrieve the geomaterials whose hardness are higher than an given value.}
\usage{
geomaterials_hardness_gt(hmin, ...)
}
\arguments{
\item{hmin}{float value of the Mohs scale of mineral hardness, which ranging from 0 to 10.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have higher hardness than an given value(hmin).
}
\details{
This function releated to the field "hardness_min" of geomaterials.
retrieve all the geomaterials that has higher hardness than the given value(hmin).
hmin:the given value of minimum Moh's hardness
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_hardness_gt(9)
}
