% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_hardness_lt}
\alias{geomaterials_hardness_lt}
\title{retrieve the geomaterials whose hardness are lower than an given value.}
\usage{
geomaterials_hardness_lt(hmax, ...)
}
\arguments{
\item{hmax}{float value of the Mohs scale of mineral hardness, which ranging from 0 to 10.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower hardness than an given vlaue(hmax).
}
\details{
This function releated to the field "hardness_max" of geomaterials.
retrieve all the geomaterials that has lower hardness than an given value(hmax).
hamx: maximum Moh's hardness
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_hardness_lt(1.5)
}
