% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ima}
\alias{geomaterials_ima}
\title{retrieve the geomaterials approved by IMA or not.}
\usage{
geomaterials_ima(btrue,...)
}
\arguments{
\item{btrue}{boolean value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials within or without the ima.
}
\details{
retrieve all the geomaterials that are approved by the IMA or not.
}
\examples{
geomaterials_ima(TRUE)
geomaterials_ima(FALSE)
}
