% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ima_notes}
\alias{geomaterials_ima_notes}
\title{retrieve the geomaterials match given notes.}
\usage{
geomaterials_ima(btrue,...)
}
\arguments{
\item{enum_item}{Ima notes: multiple choice (OR) : "GROUP" "INTERMEDIATE" "NAMED_AMPHIBOLE" "PENDING_APPROVAL" "PUBLISHED_WITHOUT_APPROVAL" "REDEFINED" "REJECTED" "RENAMED" "UNNAMED_INVALID" "UNNAMED_VALID"}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with an given .
}
\details{
retrieve all the geomaterials that match the input IMA notes.
}
\examples{
geomaterials_ima_notes("PUBLISHED_WITHOUT_APPROVAL")
}
