% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ima_status}
\alias{geomaterials_ima_status}
\title{retrieve the geomaterials matched given IMA status.}
\usage{
geomaterials_ima_status(enum_status,...)
}
\arguments{
\item{enum_status}{Ima status: multiple choice (OR):"APPROVED" "DISCREDITED" "GRANDFATHERED" "PENDING_PUBLICATION" "QUESTIONABLE"}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with an given ima status.
}
\details{
This function releated to the field "ima_status" of geomaterials.
retrieve all the geomaterials that match the input IMA notes.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_ima_status("PENDING_PUBLICATION")
}
