% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_lustretype}
\alias{geomaterials_lustretype}
\title{retrieve the geomaterials that have the given lustretype.}
\usage{
geomaterials_lustretype(types, ...)
}
\arguments{
\item{types}{string of the type name.Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials that have the given lustretype.
}
\details{
Items Enum: "Adamantine" "Dull" "Earthy" "Greasy" "Metallic" "Pearly" "Resinous" "Silky" "Sub-Adamantine" "Sub-Metallic" "Sub-Vitreous" "Vitreous" "Waxy"
}
\examples{
geomaterials_lustretype(c("Adamantine"))
geomaterials_lustretype(c("Adamantine","Dull"))
}
