% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_meteoritical_code}
\alias{geomaterials_meteoritical_code}
\title{retrieve the geomaterials matched a given string in its meteoritical code.}
\usage{
geomaterials_meteoritical_code(str_name,...)
}
\arguments{
\item{str_meteoritical_code}{Text search supporting: _ as wildcards.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with a given string matched its given meteoritical_code.
}
\details{
This function releated to the field "meteoritical_code_exists" of geomaterials.
Meteoritical code exists. Include non-empty (true) / include empty only (false).
retrieve all the geomaterials that match the input str_meteoritical_code.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_meteoritical_code("TRUE")
}
