% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_meteoritical_code_exists}
\alias{geomaterials_meteoritical_code_exists}
\title{Meteoritical code exists. Include non-empty (true) / include empty only (false)}
\usage{
geomaterials_meteoritical_code_exists(true,...)
}
\arguments{
\item{bl_exists}{boolean}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with the meteoritical_code_exists or not.
}
\details{
retrieve all the geomaterials that match the input meteoritical_code_exists or not.
}
\examples{
geomaterials_meteoritical_code_exists(TRUE)
}
