% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_not_contain_elems}
\alias{geomaterials_not_contain_elems}
\title{geomaterials that do not contain the elements}
\usage{
geomaterials_not_contain_elems (ecl_elms_vector, ...)
}
\arguments{
\item{ecl_elms_vector, }{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials list.
}
\description{
retrieve the geomaterials that do not contain any of the given elements.
}
\details{
This function releated to the field "elements_exc" of geomaterials.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_not_contain_elems(c('H','O',"Si","Al","Fe"))
}
