% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_optical2v_max}
\alias{geomaterials_optical2v_max}
\title{retrieve the geomaterials that less than the given optical 2v.}
\usage{
geomaterials_optical2v_max(lt, ...)
}
\arguments{
\item{lt}{list of the signs.Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the lower optical 2v value than the given lt.
}
\details{
This function releated to the field "optical2v_max" of geomaterials.
optical2vcalc:The calculated 2V angle minimum or average of biaxial minerals
optical2vcalc2:The calculated 2V angle maximum of biaxial minerals
optical2vmeasured:The measured 2V angle minimum or average of biaxial minerals
optical2vmeasured2:The measured 2V angle maximum of biaxial minerals
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_optical2v_max(40)
}
