% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_optical2v_min}
\alias{geomaterials_optical2v_min}
\title{retrieve the geomaterials that has higher value than the given optical 2v.}
\usage{
geomaterials_optical2v_min(gt, ...)
}
\arguments{
\item{gt}{given value of optical 2v of mineral.Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the higher optical 2v value than the given gt.
}
\details{
This function releated to the field "optical2v_mix" of geomaterials.
optical2vcalc:The calculated 2V angle minimum or average of biaxial minerals
optical2vcalc2:The calculated 2V angle maximum of biaxial minerals
optical2vmeasured:The measured 2V angle minimum or average of biaxial minerals
optical2vmeasured2:The measured 2V angle maximum of biaxial minerals
}
\examples{
\dontrun{
df <-geomaterials_optical2v_min(0.1)
}
}
