% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_optical2v_range}
\alias{geomaterials_optical2v_range}
\title{retrieve the geomaterials that has the given range of optical 2v.}
\usage{
geomaterials_optical2v_range(gt,lt,...)
}
\arguments{
\item{gt}{given value of minimum of optical 2v of mineral.Please refer to the details.}

\item{lt}{an given value of maximum of optical 2v of mineral.Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the higher optical 2v value than the given lt.
}
\details{
This function releated to the field "optical2v_min" and "optical2v_max" of geomaterials.
optical2vcalc:The calculated 2V angle minimum or average of biaxial minerals
optical2vcalc2:The calculated 2V angle maximum of biaxial minerals
optical2vmeasured:The measured 2V angle minimum or average of biaxial minerals
optical2vmeasured2:The measured 2V angle maximum of biaxial minerals
}
\examples{
\dontrun{
df <-geomaterials_optical2v_range(0.1,0.2)
}
}
