% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_opticalsign}
\alias{geomaterials_opticalsign}
\title{retrieve the geomaterials that match an given optical signs.}
\usage{
geomaterials_opticalsign(signs, ...)
}
\arguments{
\item{signs}{list of the signs(string or null).
sign for uniaxial and biaxial minerals: +;-;+/- .Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials match an given optical signs.
}
\details{
This function releated to the field "opticalsign" of geomaterials.
Optical sign: single choice (Enum): "+", "+/-", "-"
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_opticalsign("-")
}
