% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_opticaltype}
\alias{geomaterials_opticaltype}
\title{retrieve the geomaterials that have the given optical types.}
\usage{
geomaterials_opticaltype(types, ...)
}
\arguments{
\item{types}{list of the types.Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the given optical types
}
\details{
Enum: "Biaxial" "Isotropic" "Uniaxial"
}
\examples{
geomaterials_opticaltype(c("Biaxial"))
geomaterials_opticaltype(c("Biaxial", "Isotropic"))
}
