% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_polytypeof}
\alias{geomaterials_polytypeof}
\title{retrieve the geomaterials by an given id of polytype of (the id of the mineral that this record is the polytype of. )}
\usage{
geomaterials_polytypeof(ptype, ...)
}
\arguments{
\item{ptype}{integer. an mindat id of the mineral that this record is the polytype of}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials by an given id for its  polytype.
A mineral that differs from another only in the stacking of similar structural units in its atomic structure
}
\details{
This function releated to the field "polytypeof" of geomaterials.
retrieve the geomaterials with an given id of polytypeof.
}
\examples{
\dontrun{
df <-geomaterials_polytypeof(0,fields = "id,name,polytypeof")
}
}
