% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_ri_gt}
\alias{geomaterials_ri_gt}
\title{retrieve the geomaterials that refractive index higher than the given value.}
\usage{
geomaterials_ri_gt(gt, ...)
}
\arguments{
\item{gt}{float value.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the higher refractive index than the given value.
}
\details{
retrieve the geomaterials with the refractive index higher than the given value.
}
\examples{
geomaterials_ri_gt(1.6)
}
