% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ri_range}
\alias{geomaterials_ri_range}
\title{retrieve the geomaterials whose refractive index is within an given range(gt,lt).}
\usage{
geomaterials_ri_range(gt,lt, ...)
}
\arguments{
\item{gt}{float value. Refractive index, from (rimax>=).}

\item{lt}{float value. Refractive index, to (rimin<=)}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have refractive index within an given range(gt,lt).
}
\details{
This function releated to the fields "ri_min" and "ri_max" of geomaterials.
retrieve all the geomaterials that has the refractive index within the range of (gt,lt).
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_ri_range(1.7,1.8)
}
