% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_streak}
\alias{geomaterials_streak}
\title{retrieve the geomaterials that have streak.}
\usage{
geomaterials_streak(str,...)
}
\arguments{
\item{str}{string. steak}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given steak.
}
\details{
retrieve the geomaterials that has the given steak.
}
\examples{
geomaterials_streak("black")
}
