% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_without_elems}
\alias{geomaterials_without_elems}
\title{geomaterials_that do not contain the elements}
\usage{
geomaterials_without_elems(ecl_elms_vector, ...)
}
\arguments{
\item{ecl_elms_vector, }{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
retrieve the geomaterials that do nont contain the input elements.Queries the list of geological materials that
do not contain the specified elements.
}
\examples{
geomaterials_without_elems(c('H','Be'))
}
