% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomeaterials_non_utf}
\alias{geomeaterials_non_utf}
\title{retrieve the geomaterials that include non-utf mineral names or not.}
\usage{
geomeaterials_non_utf(types, ...)
}
\arguments{
\item{..., }{Further named parameters.Other optional arguments.}

\item{types}{string of the type name.Please refer to the details.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials include non-utf mineral names or not.
}
\details{
retrieve the geomaterials that contain or not contain the non-utf name.
}
\examples{
geomeaterials_non_utf(TRUE)
}
