% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomeaterials_ordering}
\alias{geomeaterials_ordering}
\title{retrieve the geomaterials by an given ordering.}
\usage{
geomeaterials_ordering(ord, ...)
}
\arguments{
\item{ord}{string Enum: "approval_year" "id" "minstats__ms_locentries" "minstats__ms_photos" "name" "updttime" "weighting".}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials by an given ordering.
}
\details{
retrieve the geomaterials by an given ordering.
}
\examples{
geomeaterials_ordering("approval_year")
geomeaterials_ordering("minstats__ms_photos")
}
