% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomeaterials_polytypeof}
\alias{geomeaterials_polytypeof}
\title{retrieve the geomaterials by an given polytypeof}
\usage{
geomeaterials_polytypeof(ptype, ...)
}
\arguments{
\item{ptype}{integer .}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials by an given polytypeof.
}
\details{
retrieve the geomaterials with an given polytypeof.
}
\examples{
geomeaterials_polytypeof("approval_year")
}
