% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_data_parse.R
\name{GET}
\alias{GET}
\title{collapse_array_columns_map}
\usage{
collapse_array_columns_map (base_uri)
}
\arguments{
\item{name}{description}
}
\value{
description
}
\description{
set up of the mindat basic uri, endpoints, and cache
}
\examples{
collapse_array_columns_map(c('id'))
}
