% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minda_data_maker.R
\name{getExtension}
\alias{getExtension}
\title{Output the file extension of a filename}
\usage{
getExtension (filename)
}
\arguments{
\item{filename}{R dataframe of retrieived data from Mindat database.}
}
\description{
Convert the mindat R dataframe to JSON-LD string
}
