% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_list_all}
\alias{localities_list_all}
\title{retrieve the localities list.}
\usage{
localities_list_all(...)
}
\arguments{
\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries the list of localities.
}
\details{
This function releated to  the fields "ids" of localities.
retrieve all the localities.
}
\examples{
\dontrun{
df <-localities_list_all(fields = "id,name,latitude,longitude")
}
}
