% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{localities_list_description}
\alias{localities_list_description}
\title{retrieve the localities that contain the given description}
\usage{
localities_list_description(desc,...)
}
\arguments{
\item{desc}{string,}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries the list of localities that contain the given description.
}
\details{
retrieve the localities that contain the given description
}
\examples{
localities_list_description("Chinese")
}
