% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_list_elems_inc}
\alias{localities_list_elems_inc}
\title{localities that contain the given elements}
\usage{
localities_list_elems_inc(inc_elems_list, ...)
}
\arguments{
\item{inc_elems_list}{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of localities
}
\description{
Queries the list of localities that contain the given elements.
}
\details{
This function releated to the field "elements_inc" of localities.
This function queries the list of localities that contain the given elements.
}
\examples{
\dontrun{
df <-localities_list_elems_inc(c("Dy"))
}
}
