% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_list_elems_inc_exc}
\alias{localities_list_elems_inc_exc}
\title{localities that contain the given elements but not contain some other given elements.}
\usage{
localities_list_elems_inc_exc(inc_elems_list,exc_elems_list, ...)
}
\arguments{
\item{inc_elems_list}{vector of elements.}

\item{exc_elems_list}{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of localities
}
\description{
Queries the list of localities that contain the given elements,but not contain some other given elements.
}
\details{
This function releated to the fields "elements_inc" and "elements_exc" of localities.
This function queries the list of localities that contain the given elements,but not contain some other given elements.
}
\examples{
\dontrun{
df <-localities_list_elems_inc_exc(c("Dy"), c("Li"))
}
}
