% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_list_txt}
\alias{localities_list_txt}
\title{localities that contain the given txt name.}
\usage{
localities_list_txt(txt,...)
}
\arguments{
\item{txt}{string.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
Queries the list of localities that contain the given txt name.
}
\details{
This function releated to the fields "txt" of localities.
This function queries the list of localities that contain the given txt name.
}
\examples{
\dontrun{
df <-localities_list_txt("lava")
}
}
