% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_retrieve_id}
\alias{localities_retrieve_id}
\title{retrieve the localities by a given mindat id.}
\usage{
localities_retrieve_id(id,...)
}
\arguments{
\item{id, }{integer. the mindat localitiy id.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries the localitiy by given id.
}
\details{
This function releated to all the fields "id" of localities.
retrieve the localities by a given id.
}
\examples{
\dontrun{
df <-localities_retrieve_id(3337)
}
}
