% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_status_list}
\alias{localities_status_list}
\title{localities_status_list}
\usage{
localities_status_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality status
}
\description{
retrieve all locality status list.
}
\details{
This function is to retrieve all the locality_status list.
}
\examples{
 localities_status_list()
 localities_status_list(fields = "id,name")
}
