% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{localitiy_type_retrieve}
\alias{localitiy_type_retrieve}
\title{localitiy_type_retrieve}
\usage{
localitiy_type_retrieve (id,...)
}
\arguments{
\item{id}{the mindat localitiy status id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality status.
}
\description{
retrieve locality type by its id.
}
\examples{
 localitiy_type_retrieve(1)
 localitiy_type_retrieve(fields = "id,name")
}
