% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{locality_age}
\alias{locality_age}
\title{locality_age}
\usage{
locality_age (id)
}
\arguments{
\item{id}{the mindat localitiy id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality age.
}
\description{
retrieve locality age by its id
}
\examples{
 locality_age(id = 222)
 locality_age(id = 222, fields = "id,name")
}
