% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{locality_age_list}
\alias{locality_age_list}
\title{locality_age_list}
\usage{
locality_age_list (...)
}
\arguments{
\item{..., }{Further named parameters.}

\item{id}{the mindat localitiy age id}
}
\value{
df, data frame of locality age.
}
\description{
retrieve all locality age list or by its conditions
}
\examples{
 locality_age_list()
 locality_age_list(fields = "id,name")
}
