% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{locality_type_list}
\alias{locality_type_list}
\title{locality_type_list}
\usage{
locality_type_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type.
}
\description{
retrieve all locality type list.
}
\examples{
 locality_type_list()
 locality_type_list(fields = "id,name")
}
