% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{Locentries_retrieve}
\alias{Locentries_retrieve}
\title{retrieve locacentries by given a Locentry ID.}
\usage{
Locentries_retrieve(...)
}
\arguments{
\item{id}{mindat Locentry id (Integer)}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries locacentries by given a ID.
}
\details{
This function return a list of locacentries.
}
\examples{
\dontrun{
df <-Locentries_retrieve(2)
}
}
