% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{Locentries_statstics_list}
\alias{Locentries_statstics_list}
\title{retrieve a full list of locacentries statstics.}
\usage{
Locentries_statstics_list(...)
}
\arguments{
\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries locacentries statstics list.
}
\details{
This function return a list of locacentries.
}
\examples{
\dontrun{
df <-Locentries_statstics_list()
}
}
