% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{mindat_api_endpoint}
\alias{mindat_api_endpoint}
\title{set up a api endpoint named name}
\usage{
mindat_api_endpoint(name,epb,ubuilder = default_uri_builder,qparams=list(),...)
}
\arguments{
\item{name}{string .}

\item{epb}{list.}

\item{ubuilder}{function default is the default_uri_builder.}

\item{qparams}{list.}

\item{...}{Further named parameters, other conditions.}
}
\value{
uri string
}
\description{
Setup up a api endpoint.
}
