% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_response_data_parse.R
\name{mindat_build_querystring}
\alias{mindat_build_querystring}
\title{mindat_build_querystring}
\usage{
mindat_build_querystring (args)
}
\arguments{
\item{args}{query args.}
}
\value{
qs. generated query string.
}
\description{
Build query string based on the query conditions.
}
\examples{
\dontrun{
mindat_cache_set('page-size',800)
ids<-c("")
hardness_min<- 9.3
fields<- c("name,hardness")
args<- list(ids,hardness_min,fields)
querystring<-mindat_build_querystring(args)
}
}
