% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\docType{data}
\name{mindat_cache}
\alias{mindat_cache}
\title{Create a cache environment named midnat_cache}
\format{
An object of class \code{environment} of length 0.
}
\usage{
mindat_cache
}
\description{
Create a new mindat cache environment.
}
\keyword{datasets}
