% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_delete}
\alias{mindat_cache_delete}
\title{Delete a cached value by the users input varname}
\usage{
mindat_cache_delete(varname)
}
\arguments{
\item{varname}{string input a cached name.Set a cached value empty by the given varname. A string, list or other objects.}
}
\value{
No return value. The cached variable named varname will be clear.
}
\description{
Remove (clear) the cache named varname in current environment.
}
\examples{
mindat_cache_delete('api_token')
}
