% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_empty}
\alias{mindat_cache_empty}
\title{Remove all cached values}
\usage{
mindat_cache_empty()
}
\value{
No return value. All cached content will be cleared.
}
\description{
Clear all current cached values. Set current environment cache empty.
}
\examples{
mindat_cache_empty()
}
