% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_get}
\alias{mindat_cache_get}
\title{Get cache value}
\usage{
mindat_cache_get(varname)
}
\arguments{
\item{varname}{string}
}
\value{
cached value. A string, list or other objects.
}
\description{
Retrieve the value of the cache named varname in current environment.
}
\examples{
token<- mindat_cache_get('api_token')
}
