% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_return_or_setup}
\alias{mindat_cache_return_or_setup}
\title{Check if the current environment has the cached function named varname.}
\usage{
mindat_cache_return_or_setup(varname,setupfun)
}
\arguments{
\item{varname}{string.}

\item{setupfun}{boolean, if the cached is a setup function.}
}
\value{
If the varname is found in current environment cache, return cached function.
If not, eval the function and return cached function.
}
\description{
Check whether the current environment has the cached function named varname,if has, return it.
if not, setup up a new cache function named varname.
}
\examples{
 aep<-  api_end_points<-mindat_cache_return_or_setup('api_end_points', function(){return (list()) })
}
