% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_set}
\alias{mindat_cache_set}
\title{Set cache name and value}
\usage{
mindat_cache_set(varname, value)
}
\arguments{
\item{varname}{string. The cached varname.}

\item{value}{string.}
}
\description{
Assigns the value to the cache named varname in current environment.
}
\examples{
mindat_cache_set("api_test", "https://api.mindat.org")
mindat_cache_set("api_token","ad9c15fa95d8063908cb5bf186c9e79f")
}
