% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_country}
\alias{mindat_country}
\title{mindat_country}
\usage{
mindat_country (id,...)
}
\arguments{
\item{id, }{country id in mindat.}

\item{..., }{Further named parameters.}
}
\value{
df, a data frame of country
}
\description{
retrieve the country by given its id.
}
\examples{
\dontrun{
 df<- mindat_country(1)
}
}
