% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_crystalclass_list}
\alias{mindat_crystalclass_list}
\title{mindat_crystalclass_list}
\usage{
mindat_crystalclass_list(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of crystalclasses list
}
\description{
retrieve all the crystalclasses list or the crystalclasses list by given conditions.
}
\examples{
\dontrun{
 df<- mindat_crystalclass_list()
}
}
