% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_crystalclasses}
\alias{mindat_crystalclasses}
\title{mindat_crystalclasses}
\usage{
mindat_crystalclasses (id,...)
}
\arguments{
\item{id}{crystalclasses id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of crystalclasses
}
\description{
retrieve crystalclasses by its id
}
\examples{
\dontrun{
 df<- mindat_crystalclasses(4)
}
}
