% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_dana8_groups}
\alias{mindat_dana8_groups}
\title{mindat_dana8_groups}
\usage{
mindat_dana8_groups(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of dana8 classification list
}
\description{
retrieve all the classifications of dana8.
}
\examples{
\dontrun{
 df<- mindat_dana8_groups()
}
}
