% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_dana8_subgroups}
\alias{mindat_dana8_subgroups}
\title{mindat_dana8_subgroups}
\usage{
mindat_dana8_subgroups(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of subgroups of dana8 classification.
}
\description{
retrieve all the subgroups of dana8.
}
\examples{
\dontrun{
 df<- mindat_dana8_subgroups()
}
}
