% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_response_data_parse.R
\name{mindat_extract_response_body}
\alias{mindat_extract_response_body}
\title{mindat_extract_response_body}
\usage{
mindat_extract_response_body (response)
}
\arguments{
\item{response}{response json}
}
\value{
if status of the response is sucess (200),return the all_data_text(the content of response).
Otherwise,report the errors.
}
\description{
.
}
\examples{
\dontrun{
library(httr)
uri<- "https://api.mindat.org/v1/geomaterials/?id_in=&hardness_min=9.3&fields=name,+
hardness&page_size=1500"
api_token<- "9ce67655d74bcd981e937be80dcea9cb"
response <- GET(uri,add_headers('Authorization'= paste('Token ',api_token,sep = "")))
raw_data <- mindat_extract_response_body(response)
}
}
