% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_geomaterial}
\alias{mindat_geomaterial}
\title{mindat_geomaterial}
\usage{
mindat_geomaterial (id)
}
\arguments{
\item{id}{geomaterial id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type list
}
\description{
retrieve  geomaterial  by its id
}
\examples{
mindat_geomaterial(id = 1)
}
