% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_geomaterial_list}
\alias{mindat_geomaterial_list}
\title{mindat_geomaterial_list}
\usage{
mindat_geomaterial_list(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type list
}
\description{
retrieve all the geomaterial list or the geomaterial by given conditions.
}
\examples{
\dontrun{
 df<- mindat_geomaterial_list()
}
}
