% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_geomaterial_search}
\alias{mindat_geomaterial_search}
\title{mindat_geomaterial_search}
\usage{
mindat_geomaterial_search (...)
}
\arguments{
\item{..., }{Further named parameters (e:Exact.If 1 returns only exact matech;ima:if 1 returns only ima-approved minerals;size:limit of returned records).}
}
\value{
df, data frame of geomaterials mathch the search
}
\description{
retrieve all the geomaterial list or the geomaterial by given conditions.
}
\examples{
\dontrun{
 df<- mindat_geomaterial_search(q="Quartz")
}
}
