% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_geomaterial_varieties}
\alias{mindat_geomaterial_varieties}
\title{mindat_geomaterial_varieties}
\usage{
mindat_geomaterial_varieties (id)
}
\arguments{
\item{id}{geomaterial id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type list
}
\description{
retrieve the geomaterial varieties by the id of geomaterial.
}
\examples{
mindat_geomaterial_varieties(id = 1)
}
