% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_localities_list}
\alias{mindat_localities_list}
\title{mindat_localities_list}
\usage{
mindat_localities_list (...)
}
\arguments{
\item{...}{Further named parameters.}
}
\value{
df. data frame of localities list.
}
\description{
retrieve localities list
}
\examples{
\dontrun{
 df<- mindat_localities_list()
}
}
