% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locality_type}
\alias{mindat_locality_type}
\title{mindat_locality_type}
\usage{
mindat_locality_type (id,...)
}
\arguments{
\item{id}{locality type id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type list
}
\description{
retrieve  locality type by its id
}
\examples{
\dontrun{
 df<- mindat_locality_type(50)
}
}
