% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locality_type_list}
\alias{mindat_locality_type_list}
\title{mindat_locality_type_list}
\usage{
mindat_locality_type_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type list
}
\description{
retrieve all locality type list
}
\examples{
\dontrun{
 df<- mindat_locality_type_list()
}
}
