% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locentries_list}
\alias{mindat_locentries_list}
\title{mindat_locentries_list}
\usage{
mindat_locentries_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locentries list.
}
\description{
retrieve Mindat locentries. A 'locentry' is a record of specific geomaterial (mineral, etc) at a specific locality.
}
\examples{
\dontrun{
 df<- mindat_locentries_list()
}
}
