% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locentries_retrieve}
\alias{mindat_locentries_retrieve}
\title{mindat_locentries_retrieve}
\usage{
mindat_locentries_retrieve (id,...)
}
\arguments{
\item{id}{Locentry id (Integer)}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locentries list.
}
\description{
retrieve Mindat locentries by a given ID. A 'locentry' is a record of specific geomaterial (mineral, etc) at a specific locality.
}
\examples{
\dontrun{
 df<- mindat_locentries_retrieve(2)
}
}
