% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locentries_stat}
\alias{mindat_locentries_stat}
\title{mindat_locentries_stat}
\usage{
mindat_locentries_stat (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locentries statstics list.
}
\description{
retrieve Mindat locality-geomaterial pairs and some statistics list.
}
\examples{
\dontrun{
 df<- mindat_locentries_stat()
}
}
