% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_mineral_ima}
\alias{mindat_mineral_ima}
\title{mindat_mineral_ima}
\usage{
mindat_mineral_ima (id,...)
}
\arguments{
\item{id}{mindat id}

\item{..., }{Further parameters.}
}
\value{
df. query resutls in data frame format.
}
\description{
retrieve ima mineral by its id.
}
\examples{
\dontrun{
 df<- mindat_mineral_ima(3337)
}
}
