% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_mineral_ima_list}
\alias{mindat_mineral_ima_list}
\title{mindat_mineral_ima_list}
\usage{
mindat_mineral_ima (...)
}
\arguments{
\item{...}{, Further named parameters.}
}
\value{
df, data frame of mineral list.
}
\description{
retrieve ima mineral list
}
\examples{
 mindat_mineral_ima_list()
 mindat_mineral_ima_list(ids = c('3','5','7','9','222'))
}
