% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_nickel_strunz10_classes}
\alias{mindat_nickel_strunz10_classes}
\title{mindat_nickel_strunz10_classes}
\usage{
mindat_nickel_strunz10_classes(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of classes of Nickel-Strunz 10th edition classifications.
}
\description{
retrieve the class list of Nickel-Strunz 10th edition classifications.
}
\examples{
\dontrun{
 df<- mindat_nickel_strunz10_classes()
}
}
