% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_nickel_strunz10_families}
\alias{mindat_nickel_strunz10_families}
\title{mindat_nickel_strunz10_families}
\usage{
mindat_nickel_strunz10_families(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of families of Nickel-Strunz 10th edition classifications.
}
\description{
retrieve the families list of Nickel-Strunz 10th edition classifications.
}
\examples{
\dontrun{
 df<- mindat_nickel_strunz10_families()
}
}
