% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_nickel_strunz10_subclasses}
\alias{mindat_nickel_strunz10_subclasses}
\title{mindat_nickel_strunz10_subclasses}
\usage{
mindat_nickel_strunz10_subclasses(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of subclasses of Nickel-Strunz 10th edition classifications.
}
\description{
retrieve the subclass list of Nickel-Strunz 10th edition classifications.
}
\examples{
\dontrun{
 df<- mindat_nickel_strunz10_subclasses()
}
}
