% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_response_data_parse.R
\name{mindat_parse_raw_data}
\alias{mindat_parse_raw_data}
\title{mindat_parse_raw_data}
\usage{
mindat_parse_raw_data (raw_data)
}
\arguments{
\item{raw_data}{content of the response body}
}
\value{
df_out, R data frame
}
\description{
parse the raw response of json to dataframe of R. If the raw_data obtained from the response is paged,
request all the pages and then add them into the df_out data frame.
}
\examples{
\dontrun{
rd<-"{\"count\":5,\"next\":null,\"previous\":null,+
\"results\":[{\"name\":\"Diamond\"},{\"name\":\"Khamrabaevite\"},+
{\"name\":\"Moissanite\"},{\"name\":\"Qingsongite\"},{\"name\":\"Uakitite\"}]}"
df<- mindat_parse_raw_data(rd)
}
}
