% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_query}
\alias{mindat_query}
\title{mindat_query}
\usage{
mindat_query (endpoint,query)
}
\arguments{
\item{endpoint}{query endpoint, e.g.'minerals_ima'.}
}
\value{
df query resutls in data frame format.
}
\description{
Basic function for query dataset at a specified endpoit.
}
\examples{
 mindat_query('minerals_ima', query = c(list(id = id)))
}
