% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_rest_api.R
\name{mindat_set_up_endpoints}
\alias{mindat_set_up_endpoints}
\title{mindat_set_up_endpoints}
\usage{
mindat_set_up_endpoints ()
}
\description{
Set up the basic mindat endpoints for current mindat API.
}
\examples{
mindat_set_up_endpoints()
}
