% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_endpoint.R
\name{mindat_setup}
\alias{mindat_setup}
\title{mindat_setup}
\usage{
mindat_setup(base_uri = 'https://api.mindat.org/v1',page_size = 800)
}
\arguments{
\item{base_uri}{base uri of mindat API.}

\item{page_size}{interger,setting the page size of responsed data from the API server.}
}
\description{
set up of the mindat basic uri, endpoints, and cache
}
\examples{
\dontrun{
mindat_setup()
}
}
