% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_spacegroups}
\alias{mindat_spacegroups}
\title{mindat_spacegroups}
\usage{
mindat_spacegroups (id,...)
}
\arguments{
\item{id}{spacegroups id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of spacegroups
}
\description{
retrieve spacegroups by its id
}
\examples{
\dontrun{
 df<- mindat_spacegroups(4)
}
}
