% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_spacegroupsets}
\alias{mindat_spacegroupsets}
\title{mindat_spacegroupsets}
\usage{
mindat_spacegroupsets (id,...)
}
\arguments{
\item{id}{spacegroupsets id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of spacegroupsets
}
\description{
retrieve spacegroups by its id
}
\examples{
\dontrun{
 df<- mindat_spacegroupsets(4)
}
}
