% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_spacegroupsets_list}
\alias{mindat_spacegroupsets_list}
\title{mindat_spacegroupsets_list}
\usage{
mindat_spacegroupsets_list(...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of spacegroupsets list
}
\description{
retrieve all the spacegroups list or the spacegroups list by given conditions.
}
\examples{
\dontrun{
 df<- mindat_spacegroups_list()
}
}
