% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_endpoint.R
\name{mindat_uri_builder}
\alias{mindat_uri_builder}
\title{mindat_uri_builder}
\usage{
mindat_uri_builder (api_base_url, config,querystring)
}
\arguments{
\item{api_base_url}{list. The base url of mindat api.}

\item{config}{list. config of current environment.}

\item{querystring}{list. list of query fields and conditions.}
}
\description{
generate the mindat_uri_builder
}
