% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindt_cache_empty}
\alias{mindt_cache_empty}
\title{Remove all cached values}
\usage{
mindt_cache_empty()
}
\description{
Clear all current cached values. Set current environment cache empty.
}
\examples{
mindt_cache_empty()
}
