% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_ima_minerals.R
\name{minerals_ima_list_expand}
\alias{minerals_ima_list_expand}
\title{minerals_ima_list_expand}
\usage{
minerals_ima_list_expand (expand,...)
}
\arguments{
\item{expand}{description}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of ima minerals with expanded fields.
}
\description{
retrieve mineral ima list with the given expand.
}
\details{
This function is related to the filed "expand" of ima mineral.
Items Enum: "~all" "*"
}
\examples{
\dontrun{
df <-minerals_ima_list_expand("~all")
}
}
