% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_ima_minerals.R
\name{minerals_ima_list_ima}
\alias{minerals_ima_list_ima}
\title{minerals_ima_list_ima}
\usage{
minerals_ima_list_ima (intValue,...)
}
\arguments{
\item{intValue}{Integer}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type.
}
\description{
retrieve mineral ima list with the given intValue.
}
\details{
This function is related to the filed "ima" of ima minerals.
Integer. 0: "PENDING_PUBLICATION"
1: "APPROVED"
}
\examples{
\dontrun{
df <-minerals_ima_list_ima(1)
}
}
