% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{minerals_ima_retrieve}
\alias{minerals_ima_retrieve}
\title{minerals_ima_retrieve}
\usage{
minerals_ima_retrieve (id,...)
}
\arguments{
\item{id}{the mindat ima id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of ima mineral by a given id.
}
\description{
retrieve mineral ima by its id.
}
\examples{
 minerals_ima_retrieve(id = 1)
}
