% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_classifications.R
\name{Nickel_strunz10_classes}
\alias{Nickel_strunz10_classes}
\title{Nickel-strunz10-classes}
\usage{
Nickel_strunz10_classes(...)
}
\arguments{
\item{..., }{Further parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of the classes of Nickel-Strunz 10th edition classifications.
}
\description{
: Queries a  list of Nickel-Strunz 10th edition classifications.
}
\details{
This function return a list of Nickel-Strunz-10 classes.
case-insensitive
}
\examples{
\dontrun{
df <-Nickel_strunz10_classes()
}
}
