% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_classifications.R
\name{Nickel_strunz10_families}
\alias{Nickel_strunz10_families}
\title{Nickel-strunz10-families}
\usage{
Nickel_strunz10_families(...)
}
\arguments{
\item{..., }{Further parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of the families of Nickel-Strunz 10th edition classifications.
}
\description{
: Queries a  list of the families of Nickel-Strunz 10th edition classifications.
}
\details{
This function return a list of Nickel-Strunz-10 families
}
\examples{
\dontrun{
df <-Nickel_strunz10_families()
}
}
