% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_classifications.R
\name{Nickel_strunz10_subclasses}
\alias{Nickel_strunz10_subclasses}
\title{Nickel-strunz10-subclasses}
\usage{
Nickel_strunz10_subclasses(...)
}
\arguments{
\item{..., }{Further parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of the subclasses of Nickel-Strunz 10th edition classifications.
}
\description{
: Queries a  list of the subclasses of Nickel-Strunz 10th edition classifications.
}
\details{
This function return a list of Nickel-Strunz-10 subclasses.
case-insensitive
}
\examples{
\dontrun{
df <-Nickel_strunz10_subclasses()
}
}
