% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{params_to_string}
\alias{params_to_string}
\title{params_to_string}
\usage{
params_to_string (params)
}
\arguments{
\item{params}{convert params to string,which is used by the mindat query function.}
}
\value{
str .
}
\description{
Prase params to string,so that the query function can deal with the other exteranl condition set by the users.
}
\examples{
\dontrun{
 params_to_string("")
}
}
