% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minda_data_maker.R
\name{saveMindatDataAs}
\alias{saveMindatDataAs}
\title{Output file as a given format}
\usage{
saveMindatDataAs (filename,format = 'csv')
}
\arguments{
\item{inputdata}{R dataframe of retrieived data from Mindat database.}

\item{outputfname}{string. the output file name.}

\item{fmt}{output format}
}
\description{
Save the mindat R dataframe to a specify format
}
\examples{
saveMindatDataAs("out.csv")
}
