% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{set_api_base}
\alias{set_api_base}
\title{set_api_base}
\usage{
set_api_base (api_base)
}
\arguments{
\item{api_base}{string. The base uri of mindat api.}
}
\description{
set base uri of current environment
}
\examples{
set_api_base("https://api.mindat.org")
}
