% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{set_api_token}
\alias{set_api_token}
\title{set_api_token}
\usage{
set_api_token (api_token)
}
\arguments{
\item{api_token}{string. The token of mindat api.}
}
\description{
set the token of current environment
}
\examples{
set_api_token("ad9c15fa95d8063908cb5bf186c9e79f")
}
