% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{set_page_size}
\alias{set_page_size}
\title{set_page_size}
\usage{
set_page_size (page_size)
}
\arguments{
\item{page_size}{string. The token of mindat api.}
}
\description{
set the page_size of response records.
}
\examples{
set_page_size(800)
}
