% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_spacegroups.R
\name{spacegroups_cclass}
\alias{spacegroups_cclass}
\title{spacegroups that match a given crystalclass ID (integer)}
\usage{
spacegroups_cclass(crystalclass_id, ...)
}
\arguments{
\item{crystalclass_id}{integer crystalclass ID .
The field "crystalclass_id" is a integer of crystalclass ID.}

\item{..., }{Further parameters like "sgtext"(space group text) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of spacegroup
}
\description{
: Queries a list of spacegroup that match a given crystalclass ID
}
\details{
This function filter data by a given given crystalclass ID.
}
\examples{
\dontrun{
df <-spacegroups_cclass(2)
}
}
