% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_spacegroups.R
\name{spacegroups_list}
\alias{spacegroups_list}
\title{return a full list of spacegroups}
\usage{
spacegroups_list(...)
}
\arguments{
\item{..., }{Further parameters like "sgtext"(space group text) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of the full list of spacegroups
}
\description{
: Queries a full list of spacegroup
}
\details{
This function return a full list of spacegroups.
}
\examples{
\dontrun{
df <-spacegroups_list()
}
}
