% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_spacegroups.R
\name{spacegroups_sgtext}
\alias{spacegroups_sgtext}
\title{spacegroups that match a given sgtext (string)}
\usage{
spacegroups_sgtext(sgtext, ...)
}
\arguments{
\item{sgtext}{string space group text (case-insensitive).
The field "sgtext" is a string of space group text.}

\item{..., }{Further parameters like "cclass"(Crystalclass) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of spacegroups
}
\description{
: Queries a list of spacegroups that match a given sgtext (string)
}
\details{
This function filter data by a given given crystalclass ID.
}
\examples{
\dontrun{
df <-spacegroups_sgtext("P1")
}
}
