% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_spacegroups.R
\name{spacegroupsets_by_id}
\alias{spacegroupsets_by_id}
\title{spacegroupsets that match a given spacegroupets ID (integer)}
\usage{
spacegroupsets_by_id(spacegroupsets_id, ...)
}
\arguments{
\item{spacegroupsets_id}{integer spacegroup ID .
The field "spacegroupets_id" is a integer of spacegroupets ID.}

\item{..., }{Further parameters like "sgtext"(space group text) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of spacegroupsets
}
\description{
: Queries a list of spacegroupsets that match a given spacegroupets ID
}
\details{
This function filter data by a given given spacegroup ID.
}
\examples{
\dontrun{
df <-spacegroupsets_by_id(2)
}
}
