% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_spacegroups.R
\name{spacegroupsets_list}
\alias{spacegroupsets_list}
\title{return a full list of spacegroupsetsets}
\usage{
spacegroupsets_list(...)
}
\arguments{
\item{..., }{Further parameters like "sgtext"(space group text) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of the full list of spacegroupsets
}
\description{
: Queries a full list of spacegroupsets
}
\details{
This function return a full list of spacegroupsets.
}
\examples{
\dontrun{
df <-spacegroupsets_list()
}
}
