% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{stop_not_param}
\alias{stop_not_param}
\title{stop_not_param}
\usage{
stop_not_param (comp_params, query)
}
\arguments{
\item{comp_params}{list. list of compulsory, the internal params in the mindat api.}

\item{query}{list of query name/value pairs.}
}
\description{
if the query param is not in the list of mindat api, stop and report the errors.Throws error if a parameter is not found in query.
This function needs to be improved. For different endpoints of mindat.org API, the fields that can be queried are different.
}
\examples{
stop_not_param(comp_params,query)
}
