% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{build_uri}
\alias{build_uri}
\title{build uri}
\usage{
build_uri(endpoint,query = list(),api_base = NULL,...)
}
\arguments{
\item{endpoint}{list.}

\item{query}{list, query conditions.}

\item{api_base}{string, base url for mindat api}

\item{...}{Further named parameters, other conditions.}
}
\value{
uri string
}
\description{
build a request uri based on .
}
