% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{build_uri}
\alias{build_uri}
\title{build uri}
\usage{
build_uri(name,endpoint_base,...,default_uri_builder,c(fields))
}
\arguments{
\item{endpoint}{list.}

\item{...}{Further named parameters, other conditions.}

\item{name}{string .}

\item{uri_builder}{function default is the default_uri_builder.}

\item{query_params}{list.}
}
\value{
uri string
}
\description{
build a request uri based on .
}
