% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_data_maker.R
\name{ConvertDF2TTL}
\alias{ConvertDF2TTL}
\title{Convert a dataframe to a string of TTL format}
\usage{
ConvertDF2TTL (inputdata,template = NULL)
}
\arguments{
\item{inputdata}{R dataframe of retrieived data from Mindat database.}

\item{template}{filepath to the template}
}
\value{
'ConvertDF2TTL()' returns a string written in TTL (pronounced 'turtle') format converted from an input R data frame (df).
}
\description{
Convert the mindat R dataframe to TTL string
}
\examples{
\dontrun{
df <-geomaterials_search_name("Quartz")
df_out <-ConvertDF2TTL(df)
}
}
