% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{default_uri_builder}
\alias{default_uri_builder}
\title{Create a default uri builder.}
\usage{
default_uri_builder(api_base_uri,config, querystring = '')
}
\arguments{
\item{api_base_uri}{string. The mindat api base uri.}

\item{config}{list of configuration.}

\item{querystring}{query string.}
}
\value{
uri string
}
\description{
Create a default uri builder which can create the request uri according to the query conditions.
}
