% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_bi_range}
\alias{geomaterials_bi_range}
\title{retrieve the geomaterials whose birifrigence are higher and lower than the given value.}
\usage{
geomaterials_bi_range(gt,lt, ...)
}
\arguments{
\item{gt}{float value.Birifrigence is calculated from refractive index as (rimax-rimin).Range: bi_min - bi_max.}

\item{lt}{float value.Birifrigence is calculated from refractive index as (rimax-rimin).Range: bi_min - bi_max.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower birifrigence than lt.
}
\details{
This function releated to the fields "bi_min"and "bi_max" of geomaterials.
retrieve all the geomaterials that has the birifrigence within the given range of (gt,lt).
}
\examples{
\dontrun{
df <-geomaterials_bi_range(0.2,0.3)
}
}
