% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_by_groupid}
\alias{geomaterials_by_groupid}
\title{retrieve the geomaterials by an given value of groupid.}
\usage{
geomaterials_by_groupid(hmin, ...)
}
\arguments{
\item{..., }{Further named parameters.Other optional arguments.}

\item{gid}{integer value}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given groupid.
}
\details{
retrieve all the geomaterials that has the given groupid.
}
\examples{
geomaterials_by_groupid(6)
geomaterials_by_groupid(3)
}
