% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_by_groupid}
\alias{geomaterials_by_groupid}
\title{retrieve the geomaterials by an given value of groupid.}
\usage{
geomaterials_by_groupid(gid,...)
}
\arguments{
\item{gid}{integer value. The id of the group to which this mineral belongs}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that match an given groupid.
}
\details{
This function releated to the field "groupid" of geomaterials.
retrieve all the geomaterials that match an given groupid.
}
\examples{
\dontrun{
df <-geomaterials_by_groupid(0)
}
}
