% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_cleavagetype}
\alias{geomaterials_cleavagetype}
\title{geomaterials that match an given cleavagetype}
\usage{
geomaterials_cleavagetype(types, ...)
}
\arguments{
\item{types}{vector of given cleavagetype (array of strings or null).
The field "cleavage" is used to describe the crystallographic orientation of cleavage directions or planes and quality.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that match an given cleavagetype
}
\details{
This function releated to the field "cleavagetype" of geomaterials.
Items Enum: "Distinct/Good" "Imperfect/Fair" "None Observed" "Perfect" "Poor/Indistinct" "Very Good"
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_cleavagetype(c('Perfect'))
}
