% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_colour}
\alias{geomaterials_colour}
\title{geomaterials that have the given colors}
\usage{
geomaterials_colour(colors, ...)
}
\arguments{
\item{colors}{vector of given colors.
colors of the mineral or rock - individual minerals at localities can also have color information.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that match a given colors.
}
\details{
This function releated to the field "colour" of geomaterials.
For example: "Brown", "Yellow", "green", "Pink","White","Orange","Blue","Gold","Dark brown","Purple".
}
\examples{
\dontrun{
df <-geomaterials_colour(c("bright blue"))
}
}
