% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_contain_any_elems}
\alias{geomaterials_contain_any_elems}
\title{geomaterials that contain any of the given elements}
\usage{
geomaterials_contain_any_elems(any_elems, ...)
}
\arguments{
\item{any_elems}{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geological materials that contain any of the specified elements.
}
\details{
This function queries the list of geological materials that contain any of the specified elements.
It performs the query operation by looping through each specified element and calling the mindat_geomaterial_list function.
}
\examples{
geomaterials_contain_any_elems(c('H','Be'))
}
