% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_dens_less_than}
\alias{geomaterials_dens_less_than}
\title{retrieve the geomaterials whose density are lower density than a given value.}
\usage{
geomaterials_dens_less_than(lt, ...)
}
\arguments{
\item{lt}{float value.
dmeas: measured density of the mineral.
This is either the lower limit (if there is a dmeas2) or average (if there is no dmeas2).
dmeas2:measured maximum density of mineral}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower density than lt.
}
\details{
This function releated to the field "density_max" of geomaterials.
retrieve all the geomaterials that has higher density than the given density(lt).
}
\examples{
\dontrun{
df <-geomaterials_dens_less_than(3)
}
}
