% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_diapheny}
\alias{geomaterials_diapheny}
\title{retrieve the geomaterials that have an given diapheny.}
\usage{
geomaterials_diapheny(diapheny, ...)
}
\arguments{
\item{diapheny}{string. The diaphany of the mineral - transparent; translucent; opaque}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have an given diapheny.
}
\details{
This function releated to the field "diapheny" of geomaterials.
The diaphany of the mineral(Items Enum): "Opaque" "Translucent" "Transparent"
}
\examples{
\dontrun{
df <-geomaterials_diapheny("Transparent")
}
}
