% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_expand}
\alias{geomaterials_expand}
\title{retrieve the geomaterials that have the given expand.}
\usage{
geomaterials_expand(expand_fields, ...)
}
\arguments{
\item{expanded_fields}{list of expand (Array of strings (Expanded fields)).Select fields to expand.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given expand.
}
\details{
This function releated to the field "expand" of geomaterials.
The field expand(Items Enum): "description" "type_localities" "localities" "relations" "~all" "*"
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_expand(c('description'))
geomaterials_expand(c('type_localities','localities'))
}
