% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_expand}
\alias{geomaterials_expand}
\title{retrieve the geomaterials that have the given expand.}
\usage{
geomaterials_expand(expanded_fields,...)
}
\arguments{
\item{expanded_fields}{list of expand (Array of strings (Expanded fields)).Select fields to expand.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given expand.
}
\details{
This function releated to the field "expand" of geomaterials.
The field expand(Items Enum): "description" "type_localities" "localities" "relations" "~all" "*"
}
\examples{
\dontrun{
df <-geomaterials_expand("~all")
}
}
