% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_field_exists}
\alias{geomaterials_field_exists}
\title{retrieve the geomaterials records of empty or not empty of a given field.}
\usage{
geomaterials_field_exists(fieldname,bexists,...)
}
\arguments{
\item{fieldname}{string}

\item{bexists}{bool}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a list of geomaterials
}
\description{
: Queries the list of geomaterials with an empty or not empty of a given field.
}
\details{
This function releated to all the fields of geomaterials.
e.g. meteoritical_code_exists.Meteoritical code exists. Include non-empty (true) / include empty only (false)
retrieve the geomaterial list with an empty or not empty of a given field.
}
\examples{
\dontrun{
df <-geomaterials_field_exists("meteoritical_code")
}
}
