% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_fracturetype}
\alias{geomaterials_fracturetype}
\title{retrieve the geomaterials that have the given fracturetype.}
\usage{
geomaterials_fracturetype(types, ...)
}
\arguments{
\item{types}{list of types.fracturetype(Array of strings or null):
How the mineral breaks-"Conchoidal" "Fibrous" "Hackly" "Irregular/Uneven" "Micaceous" "None observed" "Splintery" "Step-Like" "Sub-Conchoidal".}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given fracturetype.
}
\details{
This function releated to the field "fracturetype" of geomaterials.
fracturetype(Items Enum): "Conchoidal" "Fibrous" "Hackly" "Irregular/Uneven" "Micaceous" "None observed" "Splintery" "Step-Like" "Sub-Conchoidal"
}
\examples{
\dontrun{
df <-geomaterials_fracturetype(c("Step-Like"))
}
}
