% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_hardness_gt}
\alias{geomaterials_hardness_gt}
\title{retrieve the geomaterials whose hardness are higher than the given value.}
\usage{
geomaterials_hardness_gt(hmin, ...)
}
\arguments{
\item{hmin}{float value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have higher hardness than gt.
}
\details{
retrieve all the geomaterials that has higher hardness than the given value.
}
\examples{
geomaterials_hardness_gt(6)
geomaterials_hardness_gt(3)
}
