% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_hardness_lt}
\alias{geomaterials_hardness_lt}
\title{retrieve the geomaterials whose hardness are lower than the given value.}
\usage{
geomaterials_hardness_lt(hmax, ...)
}
\arguments{
\item{hmax}{float value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have lower hardness than hmax.
}
\details{
retrieve all the geomaterials that has lower hardness than the given value.
}
\examples{
geomaterials_hardness_lt(6)
geomaterials_hardness_lt(3)
}
