% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_hardness_range}
\alias{geomaterials_hardness_range}
\title{retrieve the geomaterials whose hardness is within the given range.}
\usage{
geomaterials_hardness_range(hmin,hmax, ...)
}
\arguments{
\item{hmin}{float value}

\item{hmax}{float value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have hardness within the given range.
}
\details{
retrieve all the geomaterials that has the hardness within the range of (hmin,hmax).
}
\examples{
geomaterials_hardness_range(3,3.5)
}
