% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_hardness_range}
\alias{geomaterials_hardness_range}
\title{retrieve the geomaterials whose hardness is within the given range.}
\usage{
geomaterials_hardness_range(hmin,hmax, ...)
}
\arguments{
\item{hmin}{float value of the Mohs scale of mineral hardness, which ranging from 0 to 10.}

\item{hmax}{float value of the Mohs scale of mineral hardness, which ranging from 0 to 10.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have hardness within the given range.
}
\details{
This function releated to the fields "hardness_min" and "hardness_max" of geomaterials.
retrieve all the geomaterials that has the hardness within an given range(hmin,hmax).
hmin:the given value of minimum Moh's hardness
hamx: maximum Moh's hardness
}
\examples{
\dontrun{
df <-geomaterials_hardness_range(2,3)
}
}
