% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ima}
\alias{geomaterials_ima}
\title{retrieve the geomaterials approved by IMA or not.}
\usage{
geomaterials_ima(btrue,...)
}
\arguments{
\item{btrue}{boolean value.TRUE IMA approved, otherwise not approved.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials within or without the ima.
}
\details{
This function releated to the field "ima" of geomaterials.
retrieve all the geomaterials that are approved by the IMA or not.
}
\examples{
library(httr)
library(jsonlite)
This_is_a_temporary_token = "2082edf7b8dab2b9887f3c2393e822c6" #You should apply for and get your own token from Mindat.org.
mindat_connection(This_is_a_temporary_token)

geomaterials_ima(TRUE)
geomaterials_ima(FALSE)
}
