% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ima_notes}
\alias{geomaterials_ima_notes}
\title{retrieve the geomaterials match given notes.}
\usage{
geomaterials_ima_notes(enum_item,...)
}
\arguments{
\item{enum_item}{Array of integers or null.
Ima notes: multiple choice (OR) : "GROUP" "INTERMEDIATE" "NAMED_AMPHIBOLE" "PENDING_APPROVAL"
"PUBLISHED_WITHOUT_APPROVAL" "REDEFINED" "REJECTED" "RENAMED" "UNNAMED_INVALID" "UNNAMED_VALID"}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with an given .
}
\details{
This function releated to the field "ima_notes" of geomaterials.
Rejected by the IMA; Pending approval; IMA Approved Group Name; Redefined by the IMA;
Renamed by the IMA; Intermediate member of a solid-solution series; Published without approval;
Unnamed (probably valid); Unnamed (probably invalid); Named Amphibole

retrieve all the geomaterials that match the input IMA notes.
}
\examples{
\dontrun{
df <-geomaterials_ima_notes(c("PENDING_APPROVAL"))
}
}
