% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ima_status}
\alias{geomaterials_ima_status}
\title{retrieve the geomaterials matched given IMA status.}
\usage{
geomaterials_ima_status(enum_status,...)
}
\arguments{
\item{enum_status}{Ima status: multiple choice (OR):"APPROVED" "DISCREDITED" "GRANDFATHERED" "PENDING_PUBLICATION" "QUESTIONABLE"}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with an given ima status.
}
\details{
This function releated to the field "ima_status" of geomaterials.
retrieve all the geomaterials that match the input IMA notes.
}
\examples{
\dontrun{
df <-geomaterials_ima_status(c("APPROVED"))
}
}
