% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_lustretype}
\alias{geomaterials_lustretype}
\title{retrieve the geomaterials that match an given lustretype.}
\usage{
geomaterials_lustretype(types, ...)
}
\arguments{
\item{types}{string of the type name (Array of strings or null).
adamantine, subadamtine, vitreous, subvitreous, resinous, waxy, greasy, silky, pearly, metallic, submetallic, dull, earthy}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials that match an given lustretype.
}
\details{
This function releated to the field "lustretype" of geomaterials.
lustretype(Items Enum): "Adamantine" "Dull" "Earthy" "Greasy" "Metallic" "Pearly" "Resinous" "Silky"
"Sub-Adamantine" "Sub-Metallic" "Sub-Vitreous" "Vitreous" "Waxy"
multiple choice (AND)
}
\examples{
\dontrun{
df <-geomaterials_lustretype(c("Adamantine"))
}
}
