% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_meteoritical_code}
\alias{geomaterials_meteoritical_code}
\title{retrieve the geomaterials matched a given string in its meteoritical code.}
\usage{
geomaterials_meteoritical_code(str_name,...)
}
\arguments{
\item{str_meteoritical_code}{Text search supporting: _ as wildcards.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with a given string matched its given meteoritical_code.
}
\details{
retrieve all the geomaterials that match the input meteoritical_code.
}
