% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_meteoritical_code}
\alias{geomaterials_meteoritical_code}
\title{retrieve the geomaterials matched a given string in its meteoritical code.}
\usage{
geomaterials_meteoritical_code(str_meteoritical_code,...)
}
\arguments{
\item{str_meteoritical_code}{boolean, meteoritical code exists. Include non-empty (TRUE) / include empty only (FALSE).}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials with a given string matched its given meteoritical_code.
}
\details{
This function releated to the field "meteoritical_code_exists" of geomaterials.
Meteoritical code exists. Include non-empty (true) / include empty only (false).
retrieve all the geomaterials that match the input str_meteoritical_code.
}
\examples{
\dontrun{
df <-geomaterials_meteoritical_code(TRUE)
}
}
