% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_not_contain_elems}
\alias{geomaterials_not_contain_elems}
\title{geomaterials that do not contain the elements}
\usage{
geomaterials_not_contain_elems (ecl_elms_vector, ...)
}
\arguments{
\item{ecl_elms_vector, }{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of geomaterials list.
}
\description{
retrieve the geomaterials that do not contain any of the given elements.
}
\details{
This function releated to the field "elements_exc" of geomaterials.
}
\examples{
\dontrun{
df <-geomaterials_not_contain_elems (c('Fe','S', 'O'), fields ="id,name,mindat_formula,elements")
}
}
