% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_opticalsign}
\alias{geomaterials_opticalsign}
\title{retrieve the geomaterials that have the given optical signs.}
\usage{
geomaterials_opticalsign(signs, ...)
}
\arguments{
\item{signs}{list of the signs.Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the given optical signs.
}
\details{
Enum: "+", "+/-", "-"
}
\examples{
geomaterials_opticalsign("+")
geomaterials_opticalsign("+/-")
}
