% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_opticalsign}
\alias{geomaterials_opticalsign}
\title{retrieve the geomaterials that match an given optical signs.}
\usage{
geomaterials_opticalsign(signs, ...)
}
\arguments{
\item{signs}{list of the signs(string or null).
sign for uniaxial and biaxial minerals: +;-;+/- .Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials match an given optical signs.
}
\details{
This function releated to the field "opticalsign" of geomaterials.
Optical sign: single choice (Enum): "+", "+/-", "-"
}
\examples{
\dontrun{
df <-geomaterials_opticalsign("-")
}
}
