% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_opticaltype}
\alias{geomaterials_opticaltype}
\title{retrieve the geomaterials that match an given optical type.}
\usage{
geomaterials_opticaltype(types, ...)
}
\arguments{
\item{types}{list of the types for the field of opticaltype. Please refer to the details.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials match an given optical type.
}
\details{
This function releated to the field "opticaltype" of geomaterials.
transparent mineral.
opticaltype(Enum) :"Biaxial" "Isotropic" "Uniaxial"
}
\examples{
\dontrun{
df <-geomaterials_opticaltype("Isotropic")
}
}
