% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ri_gt}
\alias{geomaterials_ri_gt}
\title{retrieve the geomaterials that refractive index higher than an given value(gt).}
\usage{
geomaterials_ri_gt(gt, ...)
}
\arguments{
\item{gt}{float value. Refractive index, from (rimax>=).}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the higher refractive index than an given value(gt).
}
\details{
This function releated to the field "ri_min" of geomaterials.
retrieve the geomaterials with the refractive index higher than an given value(gt).
}
\examples{
\dontrun{
df <-geomaterials_ri_gt(0.3)
}
}
