% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ri_lt}
\alias{geomaterials_ri_lt}
\title{retrieve the geomaterials that refractive index lower than the given value.}
\usage{
geomaterials_ri_gt(lt, ...)
}
\arguments{
\item{lt}{float value.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the lower refractive index than the given value.
}
\details{
retrieve the geomaterials with the refractive index lower than the given value.
}
\examples{
geomaterials_ri_lt(1.8)
}
