% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_ri_lt}
\alias{geomaterials_ri_lt}
\title{retrieve the geomaterials that refractive index lower than an given value(lt).}
\usage{
geomaterials_ri_lt(lt,...)
}
\arguments{
\item{lt}{float value. Refractive index, to (rimin<=)}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials have the lower refractive index than an given value(lt).
}
\details{
This function releated to the field "ri_max" of geomaterials.
retrieve the geomaterials with the refractive index lower than an given value(lt).
}
\examples{
\dontrun{
df <-geomaterials_ri_lt(0.5)
}
}
