% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_ri_range}
\alias{geomaterials_ri_range}
\title{retrieve the geomaterials whose refractive index is within the given range.}
\usage{
geomaterials_ri_range(gt,lt, ...)
}
\arguments{
\item{gt}{float value}

\item{lt}{float value}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have refractive index within the given range.
}
\details{
retrieve all the geomaterials that has the refractive index within the range of (gt,lt).
}
\examples{
geomaterials_ri_range(1.6,1.8)
}
