% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_synid}
\alias{geomaterials_synid}
\title{retrieve the geomaterials by given the synid.}
\usage{
geomaterials_synid(idnum,...)
}
\arguments{
\item{idnum}{integer, syn id number}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given synid.
}
\details{
retrieve the geomaterials that has the given steak.
}
\examples{
geomaterials_synid(0)
geomaterials_synid(2897)
}
