% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_synid}
\alias{geomaterials_synid}
\title{retrieve the geomaterials by an given synid.}
\usage{
geomaterials_synid(idnum,...)
}
\arguments{
\item{idnum}{integer,an given synonym id.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that match an given synid.
}
\details{
This function releated to the fields "synid" of geomaterials.
The id of the geomaterial that is the synonym of this record (this geomaterial cannot be added to a locality).
retrieve the geomaterials that has an given synid.
}
\examples{
\dontrun{
df <-geomaterials_synid(3777)
}
}
