% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_updated_at}
\alias{geomaterials_updated_at}
\title{retrieve the geomaterials updated at an given time.}
\usage{
geomaterials_updated_at(strDate,...)
}
\arguments{
\item{strDate}{string\if{html}{\out{<date-time>}}, Last updated datetime in format \%Y-\%m-\%d \%H:\%M:\%S}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that were updated at an given time
}
\details{
This function releated to the fields "updated_at" of geomaterials.
Last updated datetime in format \%Y-\%m-\%d \%H:\%M:\%S
retrieve the geomaterials that have the latest updated at the given time.
}
\examples{
\dontrun{
df <-geomaterials_updated_at("2020-2-12 10:15:12")
}
}
