% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{geomaterials_updated_at}
\alias{geomaterials_updated_at}
\title{retrieve the geomaterials updated at the given time.}
\usage{
geomaterials_updated_at(strDate,...)
}
\arguments{
\item{strDate}{string<date-time>, Last updated datetime in format %Y-%m-%d %H:%M:%S}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given synid.
}
\details{
retrieve the geomaterials that have the latest updated at the given time.
}
\examples{
geomaterials_updated_at("2022-03-09 01:13:59")
}
