% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_varietyof}
\alias{geomaterials_varietyof}
\title{retrieve the geomaterials that have the given varietyof.}
\usage{
geomaterials_varietyof(intvalue,...)
}
\arguments{
\item{intvalue}{integer,}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that have the given varietyof.
}
\details{
retrieve the geomaterials that have the given varietyof.
}
\examples{
geomaterials_varietyof(1720)
}
