% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomaterials_varietyof}
\alias{geomaterials_varietyof}
\title{retrieve the geomaterials that are varieties of an given id of geomaterials.}
\usage{
geomaterials_varietyof(intvalue,...)
}
\arguments{
\item{intvalue}{integer, id of mineral that has this variety.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the list of geomaterials that match the given varietyof.
}
\details{
This function releated to the fields "varietyof" of geomaterials.
Varieties are geomaterials that have a special distinction from the main geomaterial ie. amethyst var. quartz
retrieve the geomaterials that are varieties of an given id of geomaterials.
}
\examples{
\dontrun{
df <-geomaterials_varietyof(3337)
}
}
