% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomeaterials_non_utf}
\alias{geomeaterials_non_utf}
\title{retrieve the geomaterials that include non-utf mineral names or not.}
\usage{
geomeaterials_non_utf(btrue =TRUE,...)
}
\arguments{
\item{btrue}{boolean. Include non-UTF mineral names?.Default is TRUE.}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials include non-utf mineral names or not.
}
\details{
This function releated to the field "non_utf" of geomaterials.
retrieve the geomaterials that contain (or not contain) the non-utf name.
}
\examples{
\dontrun{
df <-geomeaterials_non_utf(TRUE,fields = "id,name,non_utf")
}
}
