% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_geomaterials.R
\name{geomeaterials_ordering}
\alias{geomeaterials_ordering}
\title{retrieve the geomaterials by an given ordering.}
\usage{
geomeaterials_ordering(ord, ...)
}
\arguments{
\item{ord}{string of field.
Prepend "-" to the field name for descending order.
Enum: "approval_year" "id" "minstats__ms_locentries" "minstats__ms_photos" "name" "updttime" "weighting".}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of geomaterials
}
\description{
: Queries the geomaterials by an given ordering.
}
\details{
This function releated to the field "ordering" of geomaterials.
ordering=-id - sort by id descending. Prepend "-" to the field name for descending order.
fields:"approval_year" "id" "minstats__ms_locentries" "minstats__ms_photos" "name" "updttime" "weighting".
retrieve the geomaterials by an given ordering.
}
\examples{
\dontrun{
df <-geomeaterials_ordering(-id)
}
}
