% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{localities_list_elems_exc}
\alias{localities_list_elems_exc}
\title{localities that do not contain the given elements}
\usage{
localities_list_elems_exc(exc_elems_list, ...)
}
\arguments{
\item{exc_elems_list}{vector of elements.}

\item{..., }{Further named parameters.Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of localities
}
\description{
Queries the list of localities that do not contain the given elements.
}
\details{
This function releated to the field "elements_exc" of localities.
This function queries the list of localities that do contain the specified elements.
}
\examples{
\dontrun{
df <-localities_list_elems_exc(c("H", "O", "Si","Fe"),fields="id,name")
}
}
