% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{localities_list_updated_at}
\alias{localities_list_updated_at}
\title{retrieve the localities list updated at the given time.}
\usage{
localities_list_updated_at(updateDate,...)
}
\arguments{
\item{updateDate}{string<date-time>, Last updated datetime in format %Y-%m-%d %H:%M:%S}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries the list of localities that have the given time
}
\details{
retrieve the localities that have the latest updated at the given time.
}
\examples{
localities_list_updated_at("2022-03-09 01:13:59")
}
