% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{locality_age}
\alias{locality_age}
\title{locality_age}
\usage{
locality_age (id,...)
}
\arguments{
\item{id, }{the mindat localitiy age id.}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality age.
}
\description{
retrieve locality age by its id
}
\details{
This function releated to the fields "id" of locality_age and locality.
}
\examples{
\dontrun{
df <-locality_age(3337)
}
}
