% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{locality_age_list}
\alias{locality_age_list}
\title{locality_age_list}
\usage{
locality_age_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality age.
}
\description{
retrieve all locality age list or by its conditions
}
\details{
This function is to retrieve all the locality_age list.
}
\examples{
\dontrun{
df <-locality_age_list()
}
}
