% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_localities.R
\name{locality_type_list}
\alias{locality_type_list}
\title{locality_type_list}
\usage{
locality_type_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type.
}
\description{
retrieve all locality type list.
}
\details{
This function is to retrieve the locality types list.
}
\examples{
\dontrun{
df <-locality_type_list()
}
}
