% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_api_endpoint}
\alias{mindat_api_endpoint}
\title{setup up a api endpoint.}
\usage{
mindat_api_endpoint(name,endpoint_base,...,default_uri_builder,c(''))
}
\arguments{
\item{name}{string .}

\item{endpoint_base}{list.}

\item{...}{other conditions.}

\item{uri_builder}{fun. e.g. fields = "id,name"}

\item{query_params}{list.}
}
\value{
uri string
}
\description{
Setup up a api endpoint.
}
\examples{
mindat_api_endpoint("geomaterials","geomaterials/\%s",default_uri_builder,c(''))
}
