% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{mindat_api_endpoint}
\alias{mindat_api_endpoint}
\title{set up a api endpoint named name}
\usage{
mindat_api_endpoint(name,endpoint_base,...,default_uri_builder,c(fields))
}
\arguments{
\item{name}{string .}

\item{endpoint_base}{list.}

\item{uri_builder}{function default is the default_uri_builder.}

\item{query_params}{list.}

\item{...}{Further named parameters, other conditions.}
}
\value{
uri string
}
\description{
Setup up a api endpoint.
}
\examples{
mindat_api_endpoint("geomaterials","geomaterials/\%s",default_uri_builder,c(''))
}
