% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_data_parse.R
\name{mindat_build_querystring}
\alias{mindat_build_querystring}
\title{mindat_build_querystring}
\usage{
mindat_build_querystring (args)
}
\arguments{
\item{args}{query args.}
}
\value{
qs. generated query string.
}
\description{
Build query string based on the query conditions.
}
\examples{
mindat_build_querystring(args)
}
