% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_has}
\alias{mindat_cache_has}
\title{Check if the current environment has the cached value of varname.}
\usage{
mindat_cache_has(varname)
}
\arguments{
\item{varname}{string.}
}
\value{
Boolean value. if the varname is found in current environment cache, return True otherwise return False.
}
\description{
Check whether or not the current environment has the cache named varname.
}
\examples{
b_has <- mindat_cache_has('api_token')
}
