% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_cache.R
\name{mindat_cache_set}
\alias{mindat_cache_set}
\title{Set cache name and value}
\usage{
mindat_cache_set(varname, value)
}
\arguments{
\item{varname}{string. The cached varname.}

\item{value}{string.}
}
\value{
No return value. The value will be cached in memory, and the cached value can be fetched by calling the function mindat_cache_get with the assigned varname.
}
\description{
Assigns the value to the cache named varname in current environment.
}
\examples{
mindat_cache_set('api_token',"9ce67655d74bcd981e937be80dcea9cb")
}
