% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_connection}
\alias{mindat_connection}
\title{Initializing Mindat API}
\usage{
mindat_connection(token, base_url = default, fmt = "json")
}
\arguments{
\item{token}{string. You can apply a token from Mindat.org.}

\item{base_url}{string.The base url of mindat API, default is "https://api.mindat.org".}

\item{fmt}{string. The format of the request and response, default is json.}
}
\description{
Initializing API Call. Setup the base_url, token and format.
}
\examples{
mindat_connection("ad9c15fa95d8063908cb5bf186c9e79f", "https://api.mindat.org","json")
mindat_connection("ad9c15fa95d8063908cb5bf186c9e79f")
}
