% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_connection}
\alias{mindat_connection}
\title{Initializing Mindat API}
\usage{
mindat_connection(token, base_url = "https://api.mindat.org",fmt ="json",page_size = 800)
}
\arguments{
\item{token}{string. You can apply a token from Mindat.org.}

\item{base_url}{string.The base url of mindat API, default is "https://api.mindat.org".}

\item{fmt}{string. The format of the request and response, default is json.}

\item{page_size}{interger, setting the page size of responsed data from the API server.}
}
\value{
No return value. A connection to the Mindat server will be established with your input token cached.
}
\description{
Initializing API Call. Setup the base_url, token and format.
}
\examples{
mindat_connection("9ce67655d74bcd981e937be80dcea9cb",page_size = 1500)
}
