% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_countries}
\alias{mindat_countries}
\title{mindat_countries}
\usage{
mindat_countries (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of countries list
}
\description{
retrieve all countries list or the contries by given conditions.
}
\examples{
\dontrun{
 df<- mindat_countries()
}
}
