% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_response_data_parse.R
\name{mindat_extract_response_body}
\alias{mindat_extract_response_body}
\title{mindat_extract_response_body}
\usage{
mindat_extract_response_body (response)
}
\arguments{
\item{response}{response json}
}
\value{
if status of the response is sucess (200),return the all_data_text(the content of response).
Otherwise,report the errors.
}
\description{
.
}
