% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_geomaterial}
\alias{mindat_geomaterial}
\title{mindat_geomaterial}
\usage{
mindat_geomaterial (id,...)
}
\arguments{
\item{id}{geomaterial id}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type list
}
\description{
retrieve  geomaterial  by its id
}
\examples{
\dontrun{
 df<- mindat_geomaterial(3337)
}
}
