% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_get_data_from_uri}
\alias{mindat_get_data_from_uri}
\title{mindat_get_data_from_uri}
\usage{
mindat_get_data_from_uri (uri)
}
\arguments{
\item{uri}{request uri}
}
\value{
df. R data frame of the request uri.
}
\description{
retrieve data from the uri.
}
\examples{
mindat_get_data_from_uri("https://api.mindat.org/geomaterials/1/")
}
