% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_response_data_parse.R
\name{mindat_get_data_from_uri}
\alias{mindat_get_data_from_uri}
\title{mindat_get_data_from_uri}
\usage{
mindat_get_data_from_uri (uri)
}
\arguments{
\item{uri}{request uri}
}
\value{
df. R data frame of the request uri.
}
\description{
retrieve data from the uri.
}
\examples{
\dontrun{
library(httr)
uri <- "https://api.mindat.org/geomaterials/?id__in=&hardness_min=9.3&fields=name,+
hardness&page_size=1500"
mindat_cache_set('api_token',"9ce67655d74bcd981e937be80dcea9cb")
df <- mindat_get_data_from_uri(uri)
}
}
