% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locality_status}
\alias{mindat_locality_status}
\title{mindat_locality_status}
\usage{
mindat_locality_status (id,...)
}
\arguments{
\item{id}{the mindat localitiy status id.}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality status
}
\description{
retrieve all locality status by its id
}
\examples{
\dontrun{
 df<- mindat_locality_status(10)
}
}
