% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_locentries_lstm_id}
\alias{mindat_locentries_lstm_id}
\title{mindat_locentries_lstm_id}
\usage{
mindat_locentries_lstm_id (lstmid,...)
}
\arguments{
\item{id}{A unique integer value identifying this locstatsmin.}

\item{..., }{Further named parameters.}
}
\value{
df, data frame of locentries statstics list.
}
\description{
retrieve Mindat locality-geomaterial pairs and some statistics list.
}
\examples{
\dontrun{
 df<- mindat_locentries_lstm_id(2)
}
}
