% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_response_data_parse.R
\name{mindat_make_data_frame}
\alias{mindat_make_data_frame}
\title{mindat_make_data_frame}
\usage{
mindat_make_data_frame (reg_list)
}
\arguments{
\item{reg_list}{response json data to list format obj.}
}
\value{
df_out, R data frame
}
\description{
convert the response json to dataframe of R
}
