% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_querys.R
\name{mindat_mineral_ima_list}
\alias{mindat_mineral_ima_list}
\title{mindat_mineral_ima_list}
\usage{
mindat_mineral_ima_list (...)
}
\arguments{
\item{...}{, Further named parameters.}
}
\value{
df, data frame of mineral list.
}
\description{
retrieve ima mineral list
}
\examples{
\dontrun{
 df<- mindat_mineral_ima_list()
}
}
