% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_setup}
\alias{mindat_setup}
\title{mindat_setup}
\usage{
mindat_set_up_endpoints (base_uri)
}
\description{
set up of the mindat basic uri, endpoints, and cache
}
\examples{
mindat_setup('https://api.mindat.org')
}
