% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_endpoint.R
\name{mindat_setup}
\alias{mindat_setup}
\title{mindat_setup}
\usage{
mindat_set_up_endpoints (base_uri)
}
\arguments{
\item{base_uri}{base uri of mindat API.}
}
\description{
set up of the mindat basic uri, endpoints, and cache
}
