% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat.R
\name{mindat_uri_builder}
\alias{mindat_uri_builder}
\title{stop_missing_param}
\usage{
stop_missing_param (compulsory_params, query)

mindat_uri_builder (api_base_url, config,querystring)
}
\arguments{
\item{api_base_url}{list. The base url of mindat api.}

\item{config}{list. config of current environment.}

\item{querystring}{list. list of query fields and conditions.}

\item{compulsory_params}{list. The internal params in the mindat api.}

\item{query}{list. list of query fields and conditions.}
}
\description{
if the query param is not in the list of mindat api, stop and report the errors.

generate the mindat_uri_builder
}
\examples{
stop_missing_param(compulsory_params,query)
mindat_uri_builder
mindat_uri_builder(api_base_url,c(''),querystring)
}
