% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_ima_minerals.R
\name{minerals_ima_list}
\alias{minerals_ima_list}
\title{minerals_ima_list}
\usage{
minerals_ima_list (...)
}
\arguments{
\item{..., }{Further named parameters.}
}
\value{
df, data frame of locality type.
}
\description{
retrieve all mineral ima list.
}
\examples{
 minerals_ima_list()
 minerals_ima_list(fields = "id,name")
}
