% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_ima_minerals.R
\name{minerals_ima_updated_at}
\alias{minerals_ima_updated_at}
\title{retrieve the mineral_ima list updated at the given time.}
\usage{
minerals_ima_updated_at(updateDate,...)
}
\arguments{
\item{updateDate}{str, Last updated datetime in format \%Y-\%m-\%d \%H:\%M:\%S}

\item{..., }{Further named parameters.Other optional arguments.}
}
\value{
df, a data frame of localities
}
\description{
: Queries the list of mineral_ima that have the given time
}
\details{
This function is related to the filed "updated_at" of ima minerals.
retrieve the localities that have the latest updated at the given time.
}
\examples{
\dontrun{
strdate<- "2023-09-13 17:36:19"
df <-minerals_ima_updated_at(strdate)
}
}
