% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat-package.R
\docType{package}
\name{OpenMindat-package}
\alias{OpenMindat}
\alias{OpenMindat-package}
\title{OpenMindat: Quickly retrieve datasets from the 'Mindat' API}
\description{
The goal of OpenMindat R package is to provide functions for users or machines to quickly and easily retrieve datasets from the mindat.org API (https://api.mindat.org/).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quexiang/OpenMindat}
  \item Report bugs at \url{https://github.com/quexiang/OpenMindat/issues}
}

}
\keyword{internal}
