% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat-package.R
\docType{package}
\name{OpenMindat-package}
\alias{OpenMindat}
\alias{OpenMindat-package}
\title{OpenMindat: Quickly Retrieve Datasets from the 'Mindat' API}
\description{
The OpenMindat R package is used to retrieve datasets from the 'mindat.org' API (\url{https://api.mindat.org/schema/redoc/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quexiang/OpenMindat}
  \item \url{https://quexiang.github.io/OpenMindat/}
  \item Report bugs at \url{https://github.com/quexiang/OpenMindat/issues}
}

}
\author{
\strong{Maintainer}: Xiang Que \email{xiangq@uidaho.edu} (\href{https://orcid.org/0000-0002-5687-8627}{ORCID})

Authors:
\itemize{
  \item Xiaogang Ma \email{max@uidaho.edu} (\href{https://orcid.org/0000-0002-9110-7369}{ORCID})
}

}
\keyword{internal}
