% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMindat-package.R
\docType{package}
\name{OpenMindat-package}
\alias{OpenMindat}
\alias{OpenMindat-package}
\title{OpenMindat: Quickly Retrieve Datasets from the 'mindat.org' API}
\description{
'Mindat' ('mindat.org') is one of the world's most widely used databases of mineral species and their distribution. Many scientists in mineralogy, geochemistry, petrology, and other Earth and planetary disciplines have been using the 'Mindat' data. Still, an open data service and the machine interface have never been fully established. To meet the overwhelming data needs, the 'Mindat' team has built an API (\url{https://api.mindat.org/schema/redoc/}) for data access.'OpenMindat' R package provides valuable functions to bridge the data highway, connecting users' data requirements to the 'Mindat' API server and assist with retrieval and initial processing to improve efficiency further and lower the barrier of data query and access to scientists. 'OpenMindat' provides friendly and extensible data retrieval functions, including the subjects of geomaterials (e.g., rocks, minerals, synonyms, variety, mixture, and commodity), localities, and the IMA (International Mineralogical Association)-approved mineral list. 'OpenMindat' R package will accelerate the process of data-intensive studies in mineral informatics and lead to more scientific discoveries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quexiang/OpenMindat}
  \item \url{https://quexiang.github.io/OpenMindat/}
  \item Report bugs at \url{https://github.com/quexiang/OpenMindat/issues}
}

}
\author{
\strong{Maintainer}: Xiang Que \email{xiangq@uidaho.edu} (\href{https://orcid.org/0000-0002-5687-8627}{ORCID})

Authors:
\itemize{
  \item Xiaogang Ma \email{max@uidaho.edu} (\href{https://orcid.org/0000-0002-9110-7369}{ORCID})
}

}
\keyword{internal}
