% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minda_data_maker.R
\name{saveMindatDataAs}
\alias{saveMindatDataAs}
\title{Output file as a given format}
\usage{
saveMindatDataAs (inputdata,outputfname)
}
\arguments{
\item{inputdata}{R dataframe of retrieived data from Mindat database.}

\item{outputfname}{string. the output file name.}
}
\description{
Save the mindat R dataframe to a specify format
}
