% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_data_maker.R
\name{saveMindatDataAs}
\alias{saveMindatDataAs}
\title{Output file as a given format}
\usage{
saveMindatDataAs (inputdata,outputfname)
}
\arguments{
\item{inputdata}{R dataframe of retrieived data from Mindat database.}

\item{outputfname}{string. the output file name.}
}
\value{
No return value.If successful, the input data frame(df) will be saved to the specified file. Otherwise, it will report an error.
}
\description{
Save the mindat R dataframe to a specify format
}
\examples{
\dontrun{
df <-geomaterials_search_name("Quartz")
saveMindatDataAs(df,"test.jsonld")
}
}
