% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{set_api_base}
\alias{set_api_base}
\title{set_api_base}
\usage{
set_api_base (api_base)
}
\arguments{
\item{api_base}{string. The base uri of mindat api.}
}
\value{
No return value. The api based url (api_base) will be cached. Users can retrieve the value by calling mindat_cache_get('api_base').
}
\description{
set base uri of current environment
}
\examples{
set_api_base("9ce67655d74bcd981e937be80dcea9cb")
}
