% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_api_tools.R
\name{set_api_token}
\alias{set_api_token}
\title{set_api_token}
\usage{
set_api_token (api_token)
}
\arguments{
\item{api_token}{string. The token of mindat api.}
}
\value{
No return value. The api_token will be cached. Users can retrieve the value by calling mindat_cache_get('api_token').
}
\description{
set the token of current environment
}
\examples{
set_api_token("9ce67655d74bcd981e937be80dcea9cb")
}
