% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindat_spacegroups.R
\name{spacegroups_by_id}
\alias{spacegroups_by_id}
\title{spacegroups that match a given spacegroup ID (integer)}
\usage{
spacegroups_by_id(spacegroup_id, ...)
}
\arguments{
\item{spacegroup_id}{integer spacegroup ID .
The field "spacegroup_id" is a integer of spacegroup ID.}

\item{..., }{Further parameters like "sgtext"(space group text) .Other optional arguments-Additional arguments.}
}
\value{
df, a data frame of spacegroup
}
\description{
: Queries a list of spacegroup that match a given spacegroup ID
}
\details{
This function filter data by a given given spacegroup ID.
}
\examples{
\dontrun{
df <-spacegroups_by_id(2)
}
}
